On the Add Content screen, make these changes 1. A single UI which will have URL: Title: Content: (With basic formatting like WYSIWYG) Collection:
If the user provides a URL, it is considered as type: Link. If they dont provide, it is considered as a note.
Remove To-do and image from here.
Image can be added as an upload button beside +Add in home page, where users will directly upload image without any other UI. And the image will be saved and displayed in the list.